Add tests on backtest for AutoRegressivePipeline
