+++ tests/test_pipeline/test_autoregressive_pipeline.py
@@ -5,11 +5,18 @@ import pandas as pd
 import pytest
 
 from etna.datasets import TSDataset
+from etna.metrics import MAE
+from etna.metrics import MetricAggregationMode
+from etna.models import CatBoostModelPerSegment
 from etna.models import LinearPerSegmentModel
+from etna.models import NaiveModel
 from etna.pipeline import AutoRegressivePipeline
 from etna.transforms import DateFlagsTransform
 from etna.transforms import LagTransform
 from etna.transforms import LinearTrendTransform
+from tests.utils import equals_with_nans
+
+DEFAULT_METRICS = [MAE(mode=MetricAggregationMode.per_segment)]
 
 
 def test_fit(example_tsds):
@@ -108,3 +115,36 @@ def test_forecast_raise_error_if_not_fitted():
     pipeline = AutoRegressivePipeline(model=LinearPerSegmentModel(), horizon=5)
     with pytest.raises(ValueError, match="AutoRegressivePipeline is not fitted!"):
         _ = pipeline.forecast()
+
+
+@pytest.mark.long
+def test_backtest_with_n_jobs(big_example_tsdf: TSDataset):
+    """Check that AutoRegressivePipeline.backtest gives the same results in case of single and multiple jobs modes."""
+    # create a pipeline
+    pipeline = AutoRegressivePipeline(
+        model=CatBoostModelPerSegment(),
+        transforms=[LagTransform(in_column="target", lags=[1, 2, 3, 4, 5], out_column="regressor_lag_feature")],
+        horizon=7,
+        step=1,
+    )
+
+    # run forecasting
+    ts1 = deepcopy(big_example_tsdf)
+    ts2 = deepcopy(big_example_tsdf)
+    pipeline_1 = deepcopy(pipeline)
+    pipeline_2 = deepcopy(pipeline)
+    _, forecast_1, _ = pipeline_1.backtest(ts=ts1, n_jobs=1, metrics=DEFAULT_METRICS)
+    _, forecast_2, _ = pipeline_2.backtest(ts=ts2, n_jobs=3, metrics=DEFAULT_METRICS)
+
+    # compare the results taking into account NaNs
+    assert equals_with_nans(forecast_1, forecast_2)
+
+
+def test_backtest_forecasts_sanity(step_ts):
+    """Check that AutoRegressivePipeline.backtest gives correct forecasts according to the simple case."""
+    ts, expected_metrics_df, expected_forecast_df = step_ts
+    pipeline = AutoRegressivePipeline(model=NaiveModel(), horizon=5, step=1)
+    metrics_df, forecast_df, _ = pipeline.backtest(ts, metrics=[MAE()], n_folds=3)
+
+    assert np.all(metrics_df.reset_index(drop=True) == expected_metrics_df)
+    assert np.all(forecast_df == expected_forecast_df)
