+++ tests/test_transforms/test_math/test_lag_transform.py
@@ -10,14 +10,7 @@ from numpy.testing import assert_almost_equal
 from etna.datasets.tsdataset import TSDataset
 from etna.transforms.math import LagTransform
 from etna.transforms.math.lags import _OneSegmentLagTransform
-
-
-def equals_with_nans(first_df: pd.DataFrame, second_df: pd.DataFrame) -> bool:
-    """Compare two dataframes with consideration NaN == NaN is true."""
-    if first_df.shape != second_df.shape:
-        return False
-    compare_result = (first_df.isna() & second_df.isna()) | (first_df == second_df)
-    return np.all(compare_result)
+from tests.utils import equals_with_nans
 
 
 @pytest.fixture
